pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "clearance-service"
version = "0.1.0"
dependencies = [
    "fastapi>=0.95.1, <1.0.0",
    "apscheduler>=3.10.1, <4.0.0",
    "pyjwt>=2.6.0, <3.0.0",
    "pymongo>=4.3.3, <5.0.0",
    "requests>=2.29.0, <3.0.0",
    "uvicorn>=0.22.0, <1.0.0",
    "auth-checker>=1.0.0, <2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.5, <7.0.0",
    "pytest-mock>=3.10.0, <4.0.0",
    "pytest-cov>=4.0.0, <5.0.0",
    "coverage[toml]>=6.2",
    "black>=23.3.0, <24.0.0",
    "mypy>=1.2.0, <2.0.0",
    "ruff==0.0.263",
    "mkdocs-material>=9.1.6, <10.0.0",
    "Pygments>=2.10.0, <3.0.0",
    "bandit>=1.7.4, <2.0.0",
    "pre-commit>=2.16.0, <3.0.0",
    "isort>=5.10.1, <6.0.0",
    "pyupgrade>=2.29.1, <3.0.0",
    "ipython>=8.13.0, <9.0.0",
]

[tool.black]
target-version = ["py310"]
line-length = 100
include = '\.pyi?$'

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F", "W", "I001"]
src = ["clearance_service"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 100

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py310"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_context = true
